Improve TestHandleSlashingEvents, refactoring
